adding DI base
