+++ core/IoC.ts
@@ -1,0 +1,7 @@
+import {Kernel} from 'inversify';
+
+const kernel = new Kernel();
+
+
+
+export const container = kernel;
