+++ package.json
@@ -47,6 +47,7 @@
     "express": "^4.13.4",
     "filewalker": "^0.1.2",
     "http-status": "^0.2.3",
+    "inversify": "^2.0.0-rc.1",
     "reflect-metadata": "^0.1.3",
     "typescript": "^1.8.10"
   },
